Revisitation of Rubamazzetto Class and project Organization
Revisitation of Rubamazzetto Class, recreation of configuration file, resolving testing problems with Player class. Adding "weight" int instance in Card class to differentiation of cards with same characteristics but different weight in specific games. Refinement of equals() and hashCode() methods in Card class. Currently the tests passed successfully are 79 of 79.